Rename ProjectionBase to Projection
